+++ src/HeyBox.Net.Rest/API/Rest/KickOutFromChannelParams.cs
@@ -5,5 +5,5 @@ namespace HeyBox.API.Rest;
 internal class KickOutFromChannelParams
 {
     [JsonPropertyName("to_user_id")]
-    public required ulong ToUserId { get; set; }
+    public required uint ToUserId { get; set; }
 }
