+++ src/HeyBox.Net.Rest/API/Rest/SendUserMessageParams.cs
@@ -17,5 +17,5 @@ internal class SendUserMessageParams
     public required string Addition { get; init; }
 
     [JsonPropertyName("to_user_id")]
-    public required ulong ToUserId { get; set; }
+    public required uint ToUserId { get; set; }
 }
